requirements.txt
kivy==2.3.1
requests>=2.32.0,<3.0
python-dotenv>=1.0.1,<2.0
httpx>=0.27.0,<0.28
